+++ src/app/(root)/page.tsx
@@ -1,152 +1,23 @@
 "use client";
 
-import SiteButton from "@/components/siteButton";
-import InfoBox from "@/components/infoBox";
-import FeaturesSection from "@/components/featuresSection";
-import MakersSection from "@/components/makersSection";
-import HopesSection from "@/components/hopesSection";
-
-import Image from "next/image";
-import Link from "next/link";
-import { landingPageText } from "@/lib/landingPageText";
+import FeaturesSection from "./featuresSection";
+import MakersSection from "./makersSection";
+import HopesSection from "./hopesSection";
+import PromisesSection from "./promisesSection";
+import HeaderSection from "./headerSection";
 
 export default function Home() {
-  const promises = landingPageText.promises.map((promise: string) => {
-    return (
-      <InfoBox aria={promise} variant="hollow" key={promise}>
-        {promise}
-      </InfoBox>
-    );
-  });
-
   return (
-    <main className="Main flex flex-grow flex-col items-center">
-      <div className="MainSection flex w-full max-w-[1000px] flex-grow justify-center">
-        <div className="MainContainer mt-28 flex -translate-x-[18vh] flex-col items-start gap-4">
-          <h1 className="MainText max-w-[700px] text-2xl font-bold text-midnight">
-            our mission: to bring simplicity, honesty, and transparency to the
-            job market.
-          </h1>
-          <div className="ButtonContainer mt-5 flex gap-8">
-            <SiteButton
-              aria="sign up"
-              size="large"
-              variant="filled"
-              colorScheme="b1"
-            >
-              sign me up!
-            </SiteButton>
-            <Link href={"/support"}>
-              <SiteButton
-                aria="support us"
-                size="large"
-                variant="filled"
-                colorScheme="e5"
-              >
-                how can i help?
-              </SiteButton>
-            </Link>
-          </div>
-        </div>
-      </div>
-      <div className="PromisesSection mt-32 flex translate-x-[30vh] flex-col flex-wrap items-end justify-end text-right">
-        <h1 className="PromisesTitle">our promises:</h1>
-        <p className="PromisesSubtitle mb-5 font-semibold italic">
-          to businesses and job-seekers
-        </p>
-        {promises}
-        <div className="ButtonContainer mt-5 flex flex-col gap-4">
-          <SiteButton
-            aria="sign up"
-            size="large"
-            variant="filled"
-            colorScheme="b4"
-            addClasses="px-10"
-          >
-            sign me up!
-          </SiteButton>
-          <Link href={"/support"}>
-            <SiteButton
-              aria="support us"
-              size="large"
-              variant="filled"
-              colorScheme="c4"
-              addClasses="px-14"
-            >
-              support us
-            </SiteButton>
-          </Link>
-          <div className="PromisePrompt flex -translate-x-[3vh] items-end">
-            <Image
-              src="/PurpleArrow.svg"
-              alt="arrow"
-              width={100}
-              height={89}
-            ></Image>
-            <p className="PromisesPrompt max-w-44 text-center text-xs text-lilac">
-              {landingPageText.arrowprompts.promise}
-            </p>
-          </div>
-        </div>
+    <div className="LandingPage flex flex-grow flex-col items-center gap-32 py-40">
+      <div className="LandingPageContainer flex w-[75%] max-w-[1600px] flex-col gap-24">
+        <HeaderSection />
+        <PromisesSection />
       </div>
       <FeaturesSection />
-      <div className="ButtonContainer mt-14 flex translate-x-[45vh] gap-8">
-        <SiteButton
-          aria="sign up"
-          size="large"
-          variant="filled"
-          colorScheme="c5"
-          addClasses="px-10"
-        >
-          get on the list
-        </SiteButton>
-        <Link href={"/support"}>
-          <SiteButton
-            aria="support us"
-            size="large"
-            variant="filled"
-            colorScheme="b2"
-            addClasses="px-14"
-          >
-            show your support
-          </SiteButton>
-        </Link>
+      <div className="LandingPageContainer flex w-[75%] max-w-[1600px] flex-col gap-24">
+        <MakersSection />
+        <HopesSection />
       </div>
-
-      <MakersSection />
-
-      <div className="ButtonContainer mb-20 mt-5 flex max-w-2xl translate-x-[30vh] flex-wrap justify-end gap-8">
-        <Link href={"/support"}>
-          <SiteButton
-            aria="help support us"
-            size="large"
-            variant="filled"
-            colorScheme="c4"
-            addClasses="px-10"
-          >
-            help this bunch of hooligans{" "}
-          </SiteButton>
-        </Link>
-        <SiteButton
-          aria="collaborate"
-          size="large"
-          variant="filled"
-          colorScheme="b4"
-          addClasses="px-14"
-        >
-          want to collaborate with this cohort?{" "}
-        </SiteButton>
-        <SiteButton
-          aria="sign up"
-          size="large"
-          variant="filled"
-          colorScheme="e6"
-        >
-          sign up!{" "}
-        </SiteButton>
-      </div>
-
-      <HopesSection />
-    </main>
+    </div>
   );
 }
