+++ src/components/navBar.tsx
@@ -17,10 +17,10 @@ export default function NavBar() {
   }
 
   return (
-    <div className="NavBar flex h-fit w-full justify-between px-16 pt-[50px]">
+    <div className="NavBar mx-auto flex h-fit w-[98%] justify-between border-b border-midnight/20 px-16 py-6">
       <Link href={"/"}>
         <Image
-          className="Logo max-w-48 cursor-pointer transition-transform duration-300 hover:scale-105"
+          className="Logo max-w-40 cursor-pointer transition-transform duration-300 hover:scale-105"
           src="/sfjs-logo.svg"
           width={229}
           height={75}
@@ -28,7 +28,7 @@ export default function NavBar() {
         />
       </Link>
 
-      <div className="NavButtonContainer flex gap-6">
+      <div className="NavButtonContainer flex items-center gap-6">
         <SiteButton
           variant="filled"
           colorScheme="b4"
