+++ src/components/siteLabel.tsx
@@ -4,7 +4,7 @@ import clsx from "clsx";
 import {
   smallShadowColors,
   SmallShadowColorOption,
-} from "@/lib/smallShadowColors";
+} from "@/lib/stylingData/smallShadowColors";
 import getRandomColorScheme from "@/utils/getRandomColorScheme";
 import Image from "next/image";
 
@@ -31,7 +31,7 @@ const SiteLabel: React.FC<LabelProps> = ({
   }
 
   const labelClasses = clsx(
-    `Label w-fit flex relative z-[1] rounded-full font-medium transition-all duration-200 ${smallShadowColors[colorScheme]} text-eggshell text-xs py-1 tracking-widest m-1`,
+    `Label w-fit py-2 flex relative z-[1] rounded-full font-medium transition-all duration-200 ${smallShadowColors[colorScheme]} text-eggshell text-xs py-1 tracking-widest m-1`,
     {
       // variant
       "px-4": variant === "display",
