+++ src/lib/stylingData/smallShadowColors.ts
